+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version 0.4 (unreleased)
-- allow editor on non-primary window
+- allow editor on non-primary window (`EditorPlugin::in_new_window(Window::default())`)
 - breaking: require `.add_plugin(EditorPlugin::new())` instead of `.add_plugin(EditorPlugin)`
 - merge `EditorState` and `Editor`
 - fix: run editor before transform propagation